Migrate typography views to use groupResourcesView
