--- src/components/groupTypographyView/groupTypographyView.view.yaml
@@ -1,82 +1,0 @@
-elementName: rvn-group-typography-view
-
-viewDataSchema:
-  type: object
-  properties:
-    searchQuery:
-      type: string
-
-propsSchema:
-  type: object
-  properties:
-    flatGroups:
-      type: array
-      description: Array of typography groups to display
-    selectedItemId:
-      type: string
-      description: ID of currently selected item
-    colorsData:
-      type: object
-      description: Colors repository data
-    fontsData:
-      type: object
-      description: Fonts repository data
-
-refs:
-  search-input:
-    eventListeners:
-      input-change:
-        handler: handleSearchInput
-  group-*:
-    eventListeners:
-      click:
-        handler: handleGroupClick
-  typography-item-*:
-    eventListeners:
-      click:
-        handler: handleTypographyItemClick
-      dblclick:
-        handler: handleTypographyItemDoubleClick
-  add-typography-button-*:
-    eventListeners:
-      click:
-        handler: handleAddTypographyClick
-
-template:
-  - 'rtgl-view w=f h=100vh g=lg style="overflow-y: scroll;"':
-    - rtgl-view g=lg w=f h=f:
-      - 'rtgl-view h=48 w=f bgc=bg bwb=xs ph=md av=c style="position: sticky; top: 0px; z-index: 1000;"':
-        - rtgl-input#search-input s=sm placeholder="Search typography..." value=${searchQuery}:
-      - $if flatGroups.length > 0:
-          - $for group in flatGroups:
-            - rtgl-view w=f ph=md:
-              - 'rtgl-view#group-${group.id} d=h w=f av=c bgc=bg cur=p style="position: sticky; top: 48px;"':
-                - $if group.isCollapsed:
-                    - rtgl-svg wh=16 svg=folderArrowRight:
-                  $else:
-                    - rtgl-svg wh=16 svg=folderArrowDown:
-                - rtgl-view d=h av=c g=md:
-                  - rtgl-svg wh=16 svg=folder:
-                  - rtgl-text: ${group.fullLabel}
-              - $if !group.isCollapsed:
-                  - rtgl-view w=f mb=md p=sm:
-                    - rtgl-view fw=w w=f g=md:
-                      - $if group.hasChildren:
-                          - $for item in group.children:
-                              - 'rtgl-view#typography-item-${item.id} w=f bw=xs cur=p style="${item.selectedStyle}"':
-                                - rvn-font-preview fontFamily=${item.fontStyle} previewText="${item.previewText}" fileId=${item.fontFileId} fontSize=${item.fontSize} lineHeight=${item.lineHeight} color=${item.color} fontWeight=${item.fontWeight} width=700 height=80 showBorder=false:
-                                - rtgl-text s=xs c=mu-fg m=sm ellipsis w="calc(700px - 2 * var(--spacing-md))": ${item.name}
-                      - 'rtgl-view#add-typography-button-${group.id} w=f d=h av=c ah=c g=md p=lg bw=xs bc=bo br=lg cur=p':
-                        - rtgl-svg svg=plus wh=24 c=pr:
-                        - rtgl-text s=md c=pr fw=500: Add Typography
-        $elif searchQuery:
-          - rtgl-view w=f h=400 av=c ah=c:
-            - rtgl-text s=lg c=mu-fg: No typography found matching "${searchQuery}"
-        $else:
-          - rtgl-view w=f h=f av=c ah=c:
-              - rtgl-text s=lg c=mu-fg: No data
-
-              - rtgl-view h=64:
-
-      - $if flatGroups.length > 0:
-          - rtgl-view h=33vh:
